Added [Readme] to fix CPANTS warning has-readme
Description on CPANTS is as follows:
The file "README" is missing from this distribution. The README provides some basic information to users prior to downloading and unpacking the distribution. Add a README to the distribution. It should contain a quick description of your module and how to install it.

I was advised that this commit would fix it on #pr-challenge irc channel.